datas.model gets Timestamp type
The Timestamp type is basically just an int64 that represents
milliseconds since the epoch.  This lets us express times of the
precision specified in our API plans as a simple 64-bit number,
which for our range queries and storage considerations is pretty
convenient.

Also has a function for parsing string timestamps and returning
a datas.model.Timestamp, of the following formats:
  2014-04-01T12:15:00.999Z
  2014-04-01 12:15:00.999
  20140401121500.999
(All interpreted as UTC whether explicit in the string or not).